Add Item view mostly implemented
The Add Item view now opens when the Add Item button is clicked on the
Inventory view.  Clicking on a list item will bring up a dialog where
the user can input an expiration date and quantity.  Currently, the
input expiration date is not parsed properly into a Date object, so any
added item's expiration date will be the current date.